Create factors of number
Asks the user to input a positive integer. Then, the for-loop checks each number between 1 and the user's input to see if there are no remainders. If there is no remainder, then the number is a factor of the user's input. The program then prints out the factors in numerical order.